+++ nuget.config
@@ -3,5 +3,15 @@
   <packageSources>
     <clear />
     <add key="nuget.org" value="https://api.nuget.org/v3/index.json" protocolVersion="3" />
+    <add key="azure-appservice" value="https://www.myget.org/F/azure-appservice/api/v3/index.json" />
   </packageSources>
+<packageSourceMapping>
+  <!-- key value for <packageSource> should match key values from <packageSources> element -->
+  <packageSource key="nuget.org">
+    <package pattern="*" />
+  </packageSource>
+  <packageSource key="azure-appservice">
+    <package pattern="Microsoft.Azure.WebJobs" />
+  </packageSource>
+</packageSourceMapping>
 </configuration>
